changed prices in float

+++ Contiki/sensors/vending_machine.h
@@ -4,7 +4,7 @@
 
 struct product {
 	uint8_t remaining_qty;
-	uint8_t price;
+	float price;
 };
 
 struct coordinate {
